Re-work shower axis sampling scheme, dialing back the cleverness by half.
The previous scheme did not account for the depth bias inherent in sampling
core positions evenly over the projected area of the surface. Now we sample
from the zenith/multiplicity distribution in the same way at all depths,
which is both simpler and reproduces the correct rates.

NB: this commit breaks compatibility with files produces with previous
versions of MuonGun.


git-svn-id: http://code.icecube.wisc.edu/svn/projects/MuonGun/trunk@103047 16731396-06f5-0310-8873-f7f720988828