Work in progress from a few months ago to implement natural-rate sampling (see #1399). Things done so far:
- Correct implementation of MUPAGE parameterizations. It turns out that the
  energy units were wrong. Durr.
- Tentative implementation of joint sampling from the energy/radius
  distribution. For BMSS (MUPAGE) this is direct; for the spline distributions
  it uses an indepencence Metropolis-Hastings sampler with BMSS as the proposal
  distribution.

The latter is unsatisifying, as the sampler is not guaranteed to converge to
the stationary distribution, nor is the correlation between the samples known.
The obvious approach of sampling first from the unconditional radial
distribution (i.e. integrated over all energies) does not work in general,
because the parameterized radial distibution does not match the integral of the
joint distribution everywhere (they were fit separately). Straight rejection
sampling is also not great, since the location of the maximum is not generally
known, nor can one guarantee that the surface is either strictly log-concave or
log-convex. More thought (or perhaps acceptance of general slowness) is needed.



git-svn-id: http://code.icecube.wisc.edu/svn/projects/MuonGun/branches/unbiased-generation@148890 16731396-06f5-0310-8873-f7f720988828